Restructure btree and block_collection files.
